+++ mutacc/cli/importing.py
@@ -3,7 +3,7 @@
 """
 
 import logging
-import pickle
+import json
 
 import click
 from pymongo.errors import WriteError as MongoWriteError
@@ -15,8 +15,11 @@ LOG = logging.getLogger(__name__)
 
 @click.command('import')
 @click.argument('case', type=click.Path(exists=True))
+@click.option('-r', '--replace',
+              is_flag=True,
+              help="If given, the case in the database will be replaced with this")
 @click.pass_context
-def importing(context, case):
+def importing(context, case, replace):
 
     """
         Import cases to the database
@@ -24,15 +27,18 @@ def importing(context, case):
     log_msg = f"Inserting case from {case} into mutaccDB"
     LOG.info(log_msg)
 
-    with open(case, "rb") as pickle_handle:
-        case = pickle.load(pickle_handle)
+    with open(case, "r") as json_handle:
+        case = json.load(json_handle)
 
     adapter = context.obj['adapter']
 
     #Check if case_id already exists in collection
-    if adapter.case_exists(case.case_id):
-        LOG.warning("case_id not unique")
-        error_msg = f"Case {case.case_id} already exists"
-        raise MongoWriteError(error_msg)
+    if adapter.case_exists(case['case']['case_id']):
+        log_msg = f"case {case['case']['case_id']} already exists"
+        LOG.warning(log_msg)
 
-    insert_entire_case(adapter, case)
+        if not replace:
+            error_msg = f"Case {case['case']['case_id']} already exists"
+            raise MongoWriteError(error_msg)
+
+    insert_entire_case(adapter, case, replace=replace)