+++ mutacc/cli/root.py
@@ -26,9 +26,10 @@ LOG = logging.getLogger(__name__)
 @click.option('--loglevel', default='INFO', type=click.Choice(LOG_LEVELS))
 @click.option('-c', '--config-file', type=click.Path(exists=True))
 @click.option('-r', '--root-dir', type=click.Path(exists=True))
+@click.option('-d', '--demo', is_flag=True)
 @click.version_option(__version__)
 @click.pass_context
-def cli(context, loglevel, config_file, root_dir):
+def cli(context, loglevel, config_file, root_dir, demo):
 
     coloredlogs.install(level = loglevel)
 
@@ -36,27 +37,38 @@ def cli(context, loglevel, config_file, root_dir):
 
     cli_config = {}
 
-    if config_file:
+    if demo:
+        host = 'localhost'
+        port = 27017
+        db_name = 'mutacc-demo'
+        username = None
+        password = None
+        root_dir = make_dir(root_dir or './mutacc_demo_root')
 
-        with open(config_file, 'r') as in_handle:
-            cli_config = yaml.load(in_handle, Loader=yaml.FullLoader)
+    else:
 
+        if config_file:
+            with open(config_file, 'r') as in_handle:
+                cli_config = yaml.load(in_handle, Loader=yaml.FullLoader)
+
+        host = cli_config.get('host') or 'localhost'
+        port = cli_config.get('port') or 27017
+        db_name = cli_config.get('database') or 'mutacc'
+        username = cli_config.get('username')
+        password = cli_config.get('password')
+        root_dir = cli_config.get('root_dir') or root_dir
+        if not root_dir:
+            LOG.warning('Please provide a root directory, through option --root-dir or in config_file')
+            context.abort()
 
     mutacc_config = {}
-    mutacc_config['host'] = cli_config.get('host') or 'localhost'
-    mutacc_config['port'] = cli_config.get('port') or 27017
-    mutacc_config['username'] = cli_config.get('username')
-    mutacc_config['password'] = cli_config.get('password')
-    mutacc_config['db_name'] = cli_config.get('database') or 'mutacc'
-
-    #Check the root_dir and add to mutacc_config
-    root_dir = cli_config.get('root_dir') or root_dir
-    if not root_dir:
-
-        LOG.warning('Please provide a root directory, through option --root-dir or in config_file')
-        context.abort()
-
+    mutacc_config['host'] = host
+    mutacc_config['port'] = port
+    mutacc_config['username'] = username
+    mutacc_config['password'] = password
+    mutacc_config['db_name'] = db_name
     mutacc_config['root_dir'] = parse_path(root_dir, file_type = 'dir')
+    mutacc_config['demo'] = demo
 
     #Create subdirectories in root, if not already created
     for dir_type in SUB_DIRS.keys():
@@ -74,7 +86,6 @@ def cli(context, loglevel, config_file, root_dir):
     mutacc_config['binaries']['seqkit'] = binaries.get('seqkit')
 
     context.obj = mutacc_config
-
 
 
 cli.add_command(extract_command)