+++ mutacc/cli/view_command.py
@@ -3,24 +3,38 @@ import json
 from pprint import pprint
 
 
-@click.command('view')
-@click.option('-c', '--case-query')
-@click.option('-v', '--variant-query')
+@click.command('cases')
+@click.option('-c', '--case-id')
 @click.pass_context
-def view_command(context, case_query, variant_query):
+def view_cases(context,case_id):
     """
-        View cases or variants in database
+        View cases in database
+    """
+
+    adapter = context.obj['adapter']
+
+    if case_id is not None:
+        results = adapter.find_case({'case_id': case_id})
+
+    else:
+        results = adapter.find_cases({})
+
+    click.echo(pprint(results))
+
+@click.command('variants')
+@click.option('-v', '--variant-id')
+@click.pass_context
+def view_variants(context, variant_id):
+    """
+        View variants in database
     """
     adapter = context.obj['adapter']
 
     results = []
-    if case_query:
-        query = json.loads(case_query)
-        results = adapter.find_cases(query)
+    if variant_id is not None:
+        results = adapter.find_variant({'display_name': variant_id})
 
-    elif variant_query:
-
-        query = json.loads(variant_query)
-        results = adapter.find_variants(query)
+    else:
+        results = adapter.find_variants({})
 
     click.echo(pprint(results))