+++ tests/cli/test_version.py
@@ -10,25 +10,27 @@ from mutacc.cli.root import cli
 
 @patch('mutacc.cli.database.mongo_adapter.get_client')
 @patch('mutacc.cli.database.MutaccAdapter')
-def test_export(mock_mutacc_adapter, mock_get_client, mock_adapter, tmpdir):
+@patch('mutacc.builds.build_version.get_md5')
+def test_export(mock_md5, mock_mutacc_adapter, mock_get_client, mock_adapter, dataset_dir, tmpdir):
 
     """
         Test export command
     """
 
     mock_mutacc_adapter.return_value = mock_adapter
-
+    mock_md5.return_value = 'md5hash'
     root_dir = str(tmpdir.mkdir("mutacc_root_test"))
 
     runner = CliRunner()
     result = runner.invoke(cli, [
         '--root-dir', root_dir,
         'db',
-        'export',
-        '--all-variants',
-        '-m', 'child',
-        '-p',
-        '-n', 'test_sample'
+        'version',
+        '-d', dataset_dir,
+        '-c', 'comment',
+        '-m'
         ])
 
     assert result.exit_code == 0
+
+