+++ tests/parse/test_parse_ped.py
@@ -1,0 +1,9 @@
+import pytest
+
+from mutacc.parse.parse_ped import parse_ped
+
+def test_parse_ped(ped_path):
+
+    family = parse_ped(ped_path)
+    assert isinstance(family, dict)
+